+++ CHANGELOG.md
@@ -51,6 +51,7 @@ https://github.com/nwnxee/unified/compare/build8193.34...HEAD
 - Player: UpdateSkyBox();
 - Player: UpdateFogColor();
 - Player: UpdateFogAmount();
+- Player: GetGameObject()
 - Regex: Match()
 
 ### Changed